refactor(lib): move tests onto separate folder + fix configuration
