+++ tdumt-lib-tests/Resources/cameras/Cameras-1.66A-MP.bin
