+++ tdumt-lib-tests/Resources/cameras/Cameras-2.00A.bin
