+++ tdumt-lib-tests/Références de test/TDUModdingLibrary.accessor
@@ -1,0 +1,2 @@
+TDUModdingLibrary.dll
+Desktop
