+++ tdumt-lib-tests/tdumt-lib-tests.csproj
@@ -1,0 +1,75 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Project DefaultTargets="Build" ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <Import Project="..\packages\NUnit.3.13.2\build\NUnit.props" Condition="Exists('..\packages\NUnit.3.13.2\build\NUnit.props')" />
+  <PropertyGroup>
+    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
+    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
+    <ProjectGuid>{1FD0A0E2-1B4C-4CF5-8702-BEF51C9B3737}</ProjectGuid>
+    <OutputType>Library</OutputType>
+    <RootNamespace>tdumtlibtests</RootNamespace>
+    <AssemblyName>tdumt-lib-tests</AssemblyName>
+    <TargetFrameworkVersion>v4.7</TargetFrameworkVersion>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
+    <DebugSymbols>true</DebugSymbols>
+    <DebugType>full</DebugType>
+    <Optimize>false</Optimize>
+    <OutputPath>bin\Debug</OutputPath>
+    <DefineConstants>DEBUG;</DefineConstants>
+    <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
+    <Optimize>true</Optimize>
+    <OutputPath>bin\Release</OutputPath>
+    <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
+  </PropertyGroup>
+  <ItemGroup>
+    <Reference Include="System" />
+    <Reference Include="nunit.framework">
+      <HintPath>..\packages\NUnit.3.13.2\lib\net45\nunit.framework.dll</HintPath>
+    </Reference>
+    <Reference Include="DjeFramework1">
+      <HintPath>..\tdumt-lib\assemblies\DjeFramework1.dll</HintPath>
+    </Reference>
+  </ItemGroup>
+  <ItemGroup>
+    <Compile Include="Common\FileTesting.cs" />
+    <Compile Include="fileformats\banks\BnkFocusOnXenonTest.cs" />
+    <Compile Include="fileformats\banks\BNKTest.cs" />
+    <Compile Include="fileformats\binaries\CamerasTest.cs" />
+    <Compile Include="fileformats\database\DBResourceTest.cs" />
+    <Compile Include="fileformats\database\DBTest.cs" />
+    <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="ToolsTest.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Common\" />
+    <Folder Include="fileformats\" />
+    <Folder Include="Properties\" />
+    <Folder Include="Références de test\" />
+    <Folder Include="Resources\" />
+  </ItemGroup>
+  <ItemGroup>
+    <Shadow Include="Références de test\TDUModdingLibrary.accessor" />
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="Resources\banks\pc\CLK_55.bnk" />
+    <EmbeddedResource Include="Resources\banks\xenon\246_GT.bnk" />
+    <EmbeddedResource Include="Resources\cameras\Cameras-1.66A-MP.bin" />
+    <EmbeddedResource Include="Resources\cameras\Cameras-2.00A.bin" />
+    <EmbeddedResource Include="Resources\TDU_Achievements.db" />
+    <EmbeddedResource Include="Resources\TDU_Achievements.fr" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\tdumt-lib\TDUModdingLibrary.csproj">
+      <Project>{EC3A4A4F-1648-4FE9-9404-68DDA741813F}</Project>
+      <Name>TDUModdingLibrary</Name>
+    </ProjectReference>
+  </ItemGroup>
+  <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
+</Project>
