--- tdumt-lib/TDUModdingLibraryTests/Références de test/TDUModdingLibrary.accessor
@@ -1,2 +1,0 @@
-TDUModdingLibrary.dll
-Desktop
